"7:25am. I slept really well tonight. Today I do not have that dull feeling in my mind.
7:30am. No emails. Yeah, forget this. It is ghost season this time of the year. Is my F# sub post still at 0 for some reason?

https://www.reddit.com/r/fsharp/comments/11yfada/implementing_the_counterfactual_regret_algorithm/

Yeah, it is still at 0. This is the power of subjects. I bet I could have done 'Implementing ChatGPT' and I'd have gotten a 100 upvotes.

This wave of ML is the shittiest one ever. I've been very wrong with how things would go. I didn't think that backprop would keep limping along so far. The only thing I need is for the AI chips to be turds that can't beat GPUs and it will complete the shit festival.

At least with my foray into webdev, I am breaking free of the shackles of languages. It is really a breath of fresh air to be able have different processes and machines communicate remotely.

https://youtu.be/C7eSobE3Ms4
Why You Should Learn C# 2023

Let me watch this guy.

...This is pretty funny.

7:55am. Let me just finish chilling and then I'll resume the next video. I want to get it out of the way today so I can get the next part done.

8:05am. https://old.reddit.com/r/ExperiencedDevs/

8:30am. https://old.reddit.com/r/ExperiencedDevs/comments/11y8hys/chatgpt_resumes_accounted_for_30_of_the_ones_we/

///

I have heard it is good for tailoring resumes (i.e. give it a super long ‘master’ resume with your entire life story, and ask it to carve out a targeted resume for a specific job ad)

I have been meaning to try it out - I do try and tailor my resume a little bit when applying for jobs, but if you’re applying for a lot of positions there’s a limit to what’s practical

///

I haven't tried this, but I really should. Nevermind this for now.

Reading this thread is pretty interesting, as those on the hiring side wants the candidates to be honest, but also want to drop them with impunity for the slightest deviation from the ideal and don't want to spent more than quarter of a minute per resume. The HR crowd is reaping the consequences of their own actions. Next is them getting replaced by AI. As shitty as these algorithms are, they can't be worse at evaluating candidates that the hiring people.

Since they aren't taking my PL project seriously, I myself literally have zero way to prove my skills in this environment, so I have no choice but to turn my hand into a bluff despite it not being weak to begin with.

8:40am. Let me start, let me start!

8:50am. Ok, I have 5m of audio, let me just record the relevant video. I've just been doing the voice over first. I am mixing up my workflow.

It feels like every since I installed the Rx Editor, my screencasting ability has exploded. It has gotten more fun.

11:55am. Let me take a break here. So far I am 12m in. It feels like I am sprinting. I'll finish this today and then move on to implementing the MC version of CFR. Then I will cover that one.

After which comes NL Holdem. On NL Holdem I'll get a chance to bring in NNs. I really want to master the dataflow between Python and .NET."