"2:30pm. Let me resume. I am into this. I need to backtrack and explain the effect of importance sampling better.
///

We'll go into the implementation and extend the PathProbs type so it takes in a third float. What this float will represent is the behavioral path probability.

Then we'll modify the path probability multiplication operator so it takes in the extra argument.

We'll also factor it out, so that we have an operator that takes in the self instance probability in addition to the behavioral instance probability.

With these two utility functions, we can move on to modifying the two learning algorithms themselves. Of course, now that we have done this, we've broken everything everywhere, but it will be easy enough to fix just by hunting down the type errors.

///

Beautiful. Having it be read out loud also helps me find out the errors in my text. It would have been great to do this while I was writing Heaven's Key.

3:30pm.

///

We'll start things off by fixing the enumerative agent. We see a type error because we are trying to pass in a tuple of 3 floats into a function that expected two so we'll just add an extra argument to the update function and it'll go away. Then, even though this is the enumerative algorithm, there is always the possibility that it might be playing against the Monte Carlo player or something like that so to get what we had previously, we'll divide by the behavioral path probability after multiplying by the opponent path probability.

Note that if we have two enumerative agents playing against each other, the behavioral path probability will always be 1 so the algorithm will be just the same as before despite these changes.

The depth first kind of traversal that the enumerative algorithm does is very similar to sampling from an uniform distribution and in the context of the Leduc game we have here, we don't need account for the probability of sampling from an uniform distribution

///

This last point confuses even me. I've thought about this a lot, but now that I am thinking about it, I am realizing that actually no, depth first traversal is not similar to sampling from an uniform distribution.

I can imagine cases where having the behavioral probability be the uniform probability would result in errors in the presence of the depth first search. It really does need to be 1.

///

We'll start things off by fixing the enumerative agent. We see a type error because we are trying to pass in a tuple of 3 floats into a function that expected two so we'll just add an extra argument to the update function and it'll go away.

Then, even though this is the enumerative algorithm, there is always the possibility that it might be playing against the Monte Carlo player, so to get a correct result in that scenario we'll divide by the behavioral path probability after multiplying by the opponent path probability.

Note that if we have two enumerative agents playing against each other, the behavioral path probability will always be 1 so the algorithm will be just the same as before despite these changes.

The depth first kind of traversal that the enumerative algorithm does is very similar to sampling from an uniform distribution, so you might think that it might be more sound to have the behavioral instance probability be uniform instead of 1, but it is possible to imagine cases where that would lead to incorrect results. It really does need to be 1.

The chance nodes however might need adjustment, not because we have broken anything in them, but because this is a good opportunity to go more into depth on what the exact difference between sampling from an uniform distribution and depth first search is. More on that later in this video, for now let us take care of the learning algorithms.

///

///

We'll fix it in the same manner that we have done previously, by adding an extra argument where it should be in the update function, and then inside we'll divide the opponent probability by the behavioral probability in order to reproject it.

Note how the update function from the previous video was wrong. Let us stop and think about it for a moment.

If we wanted to do what we intended to, which is to just have the opponent probability factor into the value estimate update and not the self probability, since we were sampling what we should have done instead is divide by the self probability.

When sampling, the probabilities we were multiplying by, in the enumerative version expilicitly become implicit, and just the regular update without any of the multiplication terms being made explicit in the program would be equivalent to multiplying by the self, the opponent and the chance node probabilities, so to get rid of the self probability we would need to divide.

Having an explicit behavioral probability like we have now makes things a lot clearer, doesn't it?

///

///

What we are doing here is extending the sample function so it does what we want, which is to take a convex combination of the provided and an uniform distribution.

///

Whops, I am forgetting to paste the transcript here, but who cares.

///

Now we can finally get to the point of what we have been doing so far. Since we've gotten the behavioral instance probability from sample, we can plug it into the loop, in a few different places there.

First it goes into the path probability.

And if this was an enumerative algorithm, we'd only want to multiply the value estimate update by the opponent probability, so clearly the path probability that we want to plug into the update values function is the one that has already taken into account the probability of sampling our own move.

By passing a behavioral path probability that has it into the update values function we can discount the probability of our own move.

Lastly, we are no longer sampling using the current policy probability, so the x minus y should be divided by the behavioral probability.

///

///

Ghostlike thought that we might need to modify the update policy function, but he forgot that we're are using the one from the enumerative agent.

Good thing we've modified it already!

We did it for a completely different reason, but we did the right thing in the end. That's pretty convenient.

///

6:10pm. Had to go for lunch. Let me resume. I am quite into this.

///

If this is confusing, think about what the update for the value estimate should be in the enumerative case. We wouldn't want our value estimate update to be implicitly multiplied by our own action's probability.

///

The undo for the speech studio seems to be broken. I'll isolate the problem later, nevermind it for now. Let me get on with screencasting.

///

Is there anything else left over to take care of? It doesn't seem like it.

In that case, let's take care of the chance nodes. Since the Play module doesn't account for path probabilities, we only need to do that in the Learn module.

///

I keep forgetting to paste things in the journo, but maybe I should just forget about it.

///

Also perhaps in the CFR project, we'll see.

We'll start things off by making another probability multiply operator, this one specifically for the chance nodes. It'll take in 3 arguments.

///

Pasting here is a grind. Nevermind it. I do not need to keep the scripts around.

7:20pm. Unused:
https://pixabay.com/music/beats-lofi-vintage-140859/
https://pixabay.com/music/beautiful-plays-beautiful-trip-142481/

Used:
https://pixabay.com/music/modern-classical-love-mellow-piano-143300/
https://pixabay.com/music/beats-puddles-in-the-sky-142581/

9:15pm. 19.5m. If I had to dub all of this, it would take up my entire day tomorrow. Let me export it to Youtube and I'll actually make it public tomorrow.

Damn, I am tired. It would have taken me 4 days to make a video like this previously. Speech synthesis is great."