"8:50am. https://www.reuters.com/investigates/special-report/us-china-tech-cables/ https://robotic.substack.com/p/behind-the-curtain-ai
Let me read these two and then I will start. I got up a bit ago and just chilling.

9:30am. Let me start. Where was I?

https://pixabay.com/music/solo-guitar-ambient-classical-guitar-144998/
https://pixabay.com/music/ambient-epic-relaxing-flute-music-144009/
https://pixabay.com/music/beats-after-party-144155/

11:20am.

```fs
let! model       = ctx.BindFormAsync<RegisterModel>()
```

Actually, I hand't realized this just extracted the body. Now that I am at this point I finally realize something extremely basic and obvious, which is that the handler is not binding to the form specifically.

12:50pm.

///

We are making good progress. We now know how to send post requests from the client and deserialize them into .NET objects that we can use on the server side. We even return an user page after a successful login.
But returning an user page only works because we are returning the HTML view from the register handler directly which already has the user data.
We aren't redirecting the user to a different page that is on the site.
In order to be able to implement that, we need to be able to extract the user data from somewhere.
As HTTP requests are stateless, we can only get that data from the request itself.
When a user makes a request, he should also be sending his user credentials inside the cookie.
Those credentials are the very ones we are creating during the registration here.
What we have to do is just set the cookie in the response header that we send back to the client, and he will be sending them to the server on every request thereafter.
At least, until he closes the browser and the cookie is wiped.

///

I have about 14m of screencasting done, and the above is another minute that I'll have to do. What I will do next is try setting the cookie manually just to demonstrate what is going on in the browser. Then I'll bring in the Auth middleware.

12:55pm. I am losing steam, and am just in a daze. Instead of ruminating, let me have breakfast here. I've earned.

Before the day is done, I'll definitely demo how the logins are done.

After I demo the auth middleware, I'll bring in the database starting with in memory one and moving to SQLite.

2:20pm. Done with breakfast and chores. Let me resume. Let me do some more screencasting for the day.

This sense of tension is just as gripping when I was programming in the past. I already feel tired.

https://pixabay.com/music/modern-classical-the-first-star-calm-relaxing-piano-solo-music-141920/

///

In its current state the login page doesn't have much to do. In a real scenario, assuming the user doesn't have a cookie and submits the form information, we'd match it against the one in the database and give the user a cookie if the username and password match.

///

Let me do the caption annotations here. Camtasia is so shitty when the file gets bigger.

5:05pm. Had to take a break.

Let me resume. Where was I?

5:35pm. I am really tired right now. I think I'll stop the screencast here for the day.

I am 20:35m in.

Right now I am feeling blocked on where to put the password. The Giraffe example is not telling me.

Let me do just a bit more.

5:40pm. Ok, 20:43.

I need to do some research on why these claims exist. I sort of assumed that Claims are arbitrary, but that can't be it.

https://www.youtube.com/results?search_query=claim+based+authentication

These is stuff by Frank Liu here.

5:45pm. Oh, Frank Liu has the first two hours of his course on Youtube.

https://youtu.be/P2Pe9kqd3gQ?t=291

Was this a part of the course I downloaded?

Maybe I should just try asking Bing for an explanation and see what he says? Or ChatGPT.

https://stackoverflow.com/questions/44301737/what-are-the-urls-for-in-claim-types

> Known claimtypes are automatically deserialized into the context. Like http://schemas.microsoft.com/ws/2008/06/identity/claims/role is added as role to the user.roles collection (used for IsInRole).

https://www.yogihosting.com/aspnet-core-identity-claims/

6:15pm. Ok, I figured out how I want to approach the explanation. Let me have lunch here.

I'll continue the screencast tomorrow. This is not that complicated.

On one hand, screencasting is slowing me down, but on the other it is what is giving me the strength to go forward."