"9:50am. I went up staying too late browsing 4chan. Sigh, I just didn't feel like going to bed. Even playing games would have been too much of a chore.
I think I am addicted to the unpredictability of message boards. I even started participating in the recent /g/ AI discussions myself.

10:20am. Let me start. I've been slacking too much.

10:35am.

///

Welcome to Ghostlike's channel. I am your AI assistant for the day, Vilo. In the previous video we created a basic login page and thoroughly covered the path leading up to it.
Now it's time for part two.
We had time to think what we want to do in this video, and our main goal should not be to move to ASP.NET Core Identity or to bring in the fancier databases than the in memory one that we are using, but instead to set up third party authentication using the Microsoft Identity Platform.
So that will be the main goal of this video. We'll clean up after that.
Let us get started.

///

This should serve as an ice breaker.

10:40am.

```fs
            do! ctx.ChallengeAsync()
            return! redirectTo false "/login" next ctx
```

First of all, I want to investigate this challenge async. Should I be using this here or is it overwritten by redirectTo.

```cs
    /// <summary>
    /// Returns a temporary redirect response (HTTP 302) to the client.
    /// </summary>
    /// <param name="location">The URL to redirect the client to. This must be properly encoded for use in http headers
    /// where only ASCII characters are allowed.</param>
    public virtual void Redirect(string location) => Redirect(location, permanent: false);
```

It sets a status code.

10:45am. Oh, I got the slack invite. Right now I am checking mail.

///

I've been doing F# webdev tutorials on Youtube, so I just wanted to plug them here in case anybody is interested. Here is the latest one from the webdev playlist. Hope somebody finds it useful.

https://youtu.be/wQEIEDbABFU
Authentication & Authorization With ASP.NET Core And Giraffe. How To Build A Login Page. (Pt. 1)

///

10:55am. Posted this in the F# Slack.

https://stackoverflow.com/questions/42883946/forbidasync-vs-challengeasync-why-and-when-to-use-them

```fs
    /// <summary>
    /// Challenges a client to authenticate via a specific authScheme.
    /// </summary>
    /// <param name="authScheme">The name of an authentication scheme from your application.</param>
    /// <param name="next"></param>
    /// <param name="ctx"></param>
    /// <returns>A Giraffe <see cref="HttpHandler"/> function which can be composed into a bigger web application.</returns>
    let challenge (authScheme : string) : HttpHandler =
        fun (next : HttpFunc) (ctx : HttpContext) ->
            task {
                do! ctx.ChallengeAsync authScheme
                return! next ctx
            }
```

Hmmm...

Let me try using this.

11:40am. Had to take a break, let me resume.

12:10pm. https://pixabay.com/music/corporate-ambient-piano-loop-144861/

12:25pm. > We'll start work on setting up passwordless logins using the Microsoft Identity Platform.

Let me stop here for breakfast.

Currently I am 2:35m in.

I need to do some research on how to setup Azure AD. I have no idea how to proceed from here.

1pm. https://youtu.be/TAJZs1pwCOE
Secure .NET Web Apps using the Microsoft Identity Platform

1:30pm. Done with breakfast and chores. Let me resume. I watched the thing above, and they implemented logins for the default Blazor project using Azure AD.

I need to immitate the guy and get it working once myself. Let me create a fresh project and I will play with it for a bit.

1:35pm. Sigh, the video is completely different than the default templates. Let me follow it exactly.

https://youtu.be/TAJZs1pwCOE?t=38

Oh, the single org is what I need.

1:55pm. Focus me, get this done.

...Oh wait, I just see Rider is offering me to register automatically.

2pm. Ah, let me take another break here. I need it.

2:20pm. Let me resume.

But I am wondering, if this is the way to do it, what is the point of MSAL?

2:35pm. I got it to work on the template.

https://youtu.be/RQ9DKBViWEg
A new tool to help you configure Azure AD for your apps

Let me watch this.

> Code first app registration.

2:50pm.

> https://login.microsoftonline.com/ec88369d-6c2f-4f15-b0c7-adbe35caec77/oauth2/v2.0/authorize?client_id=7679d9a0-4b8b-4d19-a519-1b3b0e5ba644&redirect_uri=https%3A%2F%2Flocalhost%3A44384%2Fsignin-oidc&response_type=id_token&scope=openid%20profile&response_mode=form_post&nonce=638167279337567115.MmI1OTQ3MjktNzRlYS00NGUxLWIzOWMtZTE5OGQ5M2RhMjFjNmZmNGZjODYtNTViMy00OTFjLTljMzYtMTQ2OGM2MmE2NTli&client_info=1&x-client-brkrver=IDWeb.1.16.0.0&state=CfDJ8Onu-nldwMlLrJ_8s2SHVTSS-81o8X0Yj6qJTW2gExn4ge0GsstPrp-JAbzei0F9iaiiQLp1r6_PbWLLeucjfxnWc64mO4IdFgtWiXoAvDY5R-zuzJIB5-PF5vMRins1KjIHfrsCte5C6tgy6nW1r7qthoNVNoiC4oE56RJ1D0-aikcT6hUbUqCS41-ohjHwMIoXk3CzWhi66re3gtO80_tMucNGLck1OIZXZnkQIb-p6evuXkc7bOFqcgwLsNPyxN-NKp06sA0J-vSNy6scrZKZp56kx4-spq5uFiCOV7Iv&x-client-SKU=ID_NETSTANDARD2_0&x-client-ver=6.15.1.0&sso_reload=true

What a long URL. But it does have a return URI. Ok, I see. The callback uri is appended to it.

3pm. What do I do next?

Just give it a try.

3:10pm. Let me give it a try. I'll see how far my approach gets me.

My courage will lead me to the truth in the end.

3:55pm. This is really a huge pain in the ass. I feel completely stuck.

Let me draw out a diagram first.

4:05pm. It is difficult. I understand cookies, but I don't think I understand token based authentication.

https://youtu.be/t18YB3xDfXI
An Illustrated Guide to OAuth and OpenID Connect

https://youtu.be/t18YB3xDfXI?t=122

This is kind of nice.

Right now I am confused as to how to do basic things like get the username and the email from the user. Just logging in is not enough.

Nevermind, fuck these crappy videos.

Let me close Camtasia and everything else. This is the kind of path that I need to travel on my own. Otherwise it will drain all my will to do it.

4:25pm. Let me open Camtasia again. I'll just screencast. I do not have the motivation to do it, unless I am making content.

But let me take a break.

4:55pm. Let me try it again. I've time to think about it.

6pm.

///

As far as we can tell, the domain doesn't have to be set, but the instance does.

We've tried setting it to
http://www.google.com
and it worked so we guess that any valid url can be put in here.

/signin-oidc will get appended to the return URI, so the callback path is important.

///

Let me back this up here.

6:30pm. Done with lunch. I had a lot of inspiration in the last hour.

I figure out how to do 3rd party authentication.

ChallengeAsync.

I was wondering how the hell I should send a request to the MS site, but most likely all I need to do is call challenge async.

Camtasia crashed.

And again.

I just want to get through this section and then I will close.

7:40pm. I am tired. Let me close here for the day. I am 9m in. But I am feeling a lot more confident about my prospects now.

I should be able to finish past 2 tomorrow. Given how it started, it is amazing that I managed 9m today.

I have a blank area that I'll need to get back to. and the sound goes up to 9:30m so I am not sure what the exact amount should be, but 9m seems decent as an estimate.

7:45pm. Sigh, I should realize that ChallengeAsync is the answer to redirecting to the MS page a lot sooner. If it turns out that it is not, I will be very perplexed."