"10:25pm. https://old.reddit.com/r/ExperiencedDevs/comments/12f1hjg/should_i_let_my_company_use_my_project/
///

NO.
You go to a lawyer. You get a contract immediately where you're entirely in control. Even at that what I am saying is sketchy since I've been in meetings where the phrase "acquisition through employment" was used, after I had an issue with one of my projects. Don't even dare make a commit to that project right now while you're working until you see a lawyer and figure out what if they own that work because of a conflict.

I've walked this road. DO NOT DO IT. I'm telling you from real experience here. Just do not do it. It might be a trust me bro situation, but you have no idea how bad this can get.

That goes for the rest of you too. Do not make some cute project on the side that you plan to profit from while not reviewing your contract.

///

https://www.reddit.com/r/digitalnomad/?utm_source=reddit&utm_medium=usertext&utm_name=ExperiencedDevs&utm_content=t5_e0gez

Job sites. Forget those.

10:55am.

///

In the realm of code, where the algorithms play,
A language emerged, as the dawn of new day;
F sharp it was called, a language of grace,
With functional charm, it took its place.

Amidst the vast .NET framework's expanse,
F sharp's melody dared to enhance;
Its patterns matched, and types inferred,
A harmony of logic, elegantly conferred.

Oh, hark! The brevity of syntax divine,
In the lines of code, concision did shine;
Expressiveness ruled, as with quill on parchment,
The prose of programmers, now more enchanting.

The language's power, from functions arose,
Immutable truths, it did disclose;
Higher-order thoughts, composed and refined,
In the minds of developers, forever enshrined.

For in this land of bytes and bits,
The fusion of paradigms permits
A union of objects and functions, entwined,
In the elegant dance of F sharp's design.

Thus, let us raise our voices and declare,
The praise of F sharp, beyond compare;
For in this language, beauty and might,
A new era of code, ignites the night.

///

F# poem by ChatGPT in the style of Lord Byron. I found it on the F# slack.

4/11/2023

9:30am. Let me chill just a bit more and then I will start.

10:20am. Let me start. Where was I?

1pm. I did a fantastic job. Now I just need to explain that middle section and do the outro, and I'll be done with the video. Let me have breakfast here.

3pm. Wow, I sure wasted a lot of time doing chores. Let me resume.

3:35pm. Focus me, stop reading HN. Finish the video.

https://huggingface.co/spaces/stabilityai/stable-diffusion

This is a SD 2.1 space with Google TPUs. Since I want a thumbnail image, it is easier for me to use this than to mess with Paperspace.

It has native 768 images so I don't need to upscale. Not bad.

4:15pm. https://youtu.be/hlC6baPjFOs
Authentication & Authorization With ASP.NET Core And Giraffe. Azure AD. (Pt. 2)

Here is my latest masterpiece.

4:25pm. https://twitter.com/Ghostlike

Posted in my Twitter profile, but I'll only post my work in the F# slack occasionally.

I wonder if I should post it on LinkedIn?

Sigh, who is going to care about this social media.

4:35pm. If the viewers come let them come. If not, I'll move to contracting.

Now that I am finally done with that video, I do not feel like starting right away with the next. That is how it goes usually.

I am going to get the databases and MS Identity and Integration out of the way.

5:10pm. Let me start the next video.

5:30pm. Let me really start instead of surfing /g/.

I'll get into it once I do.

///

In this video we'll be covering the ASP.NET Core Identity library, which is distinct from the online service called the Microsoft Identity Platform.
We will also be bringing in fancier servers

///

I am not sure whether I want this in the current video. Let me cut it out.

5:55pm. I've started the video and did the intro. Let me set up the project.

5:55pm. Actually, let me just think a little of how I want to do things. I think I'll get rid of the usernames.

6:10pm. I had time to think about it, but I really do not feel like doing any more programming for the day.

6:15pm. Let me go get lunch.

Tomorrow I am going to take care of the integration. After that, I'll bring in those libs, and highlight the Giraffe examples.

After that I'll do what I want, which is work on node based UIs."