"8:55am. Let me chill for a bit and then I will start.
9:25am. Let me start. Where was I? I am going to finish part 4 today.

Camtasia is so poorly optimized. It takes forever to load. It might be worth considering Active Presenter again. I wrote it off back then, but my workflow is far different now.

9:35am. Let me start from the SO post.

///

are aiming for 100% clearance in the game we are playing here.
ChatGPT turned out to be worse than useless. Had its answer been not wholy made up, it would have been great, but as it is, I don't think we'll be using it as a source of advice again.

///

Let me cut this out.

10:10am. https://pixabay.com/music/solo-guitar-illusion-feel-ambient-guitar-146100/

10:30am.

Nope, it doesn't seem we can put anything into the middleware here. In the worst case, we'll just bring back auth_all, but is there a way to have the auth middleware do that work for use instead?

```fs
    let login : HttpHandler = fun next ctx -> task {
        let! _ = ctx.AuthenticateAsync("AzureOIDC")
```

Let me add this line to see if it does something.

11:30am. https://chsakell.com/2019/07/28/asp-net-core-identity-series-external-provider-authentication-registration-strategy/#:~:text=The%20steps%20to%20enable%20authentication,external%20provider%20in%20Startup%2FConfigureServices

Goddamit, I am so bogged down with this shit.

11:40am. https://stackoverflow.com/questions/48836688/what-exactly-is-useauthentication-for

```cs
var defaultAuthenticate = await Schemes.GetDefaultAuthenticateSchemeAsync();
if (defaultAuthenticate != null)
{
    var result = await context.AuthenticateAsync(defaultAuthenticate.Name);
    if (result?.Principal != null)
    {
        context.User = result.Principal;
    }
}
```

Why does it just try to authenticate the default? Shouldn't it try all of them?

The decompiler is useless.

12:25pm. > In ASP.NET Core 7, how do I get a list of all the authentication schema names that have been added?

This is such pain in the ass, and I am starting to hate ASP.NET Core all over again.

It wasted a shitload of my time for nothing.

Let me ask ChatGPT how to do it as I have no other option.

> My ASP.NET Core 7 web application has multiple authentication schemas. How can I set it up so it tries authenticating using all of them?

Let me try the first question, first.

> Is it possible to set up the ASP.NET 7 web app so it tries out all of the authetication schemas at the start?

```cs
using Microsoft.AspNetCore.Authentication;

public class Startup
{
    public void ConfigureServices(IServiceCollection services)
    {
        services.AddAuthentication(options =>
        {
            options.DefaultScheme = CookieAuthenticationDefaults.AuthenticationScheme;
            options.DefaultChallengeScheme = "YourOtherScheme";
        })
        .AddCookie()
        .AddJwtBearer()
        .AddOpenIdConnect("YourOtherScheme", options =>
        {
            // Configuration for the other scheme
        });

        // Other service configurations
    }

    // Other methods
}
```

ChatGPT is making making me think about some of the things that I might've overlooked.

```cs
    builder.Services.AddAuthentication(fun opt ->
        opt.DefaultScheme <- Cookie.Onsite
        opt.DefaultChallengeScheme <- "AzureOID"
        )
```

Would this enable the UseAuthentication to work?

12:35pm. ChatGPT is wrong again, it seems. It isn't authing AzureOID. I also tried Cookie.Azure and am getting nothing.

12:45pm. Let me take a break here.

It seems there are days just like these.

I think I have an understanding how to deal with this next part.

3pm. Done with chores and breakfast. Let me reusme.

```fs
    app.Use(Func<HttpContext,RequestDelegate,Threading.Tasks.Task>(fun ctx next -> task {
        match ctx.Request.Cookies.TryGetValue("auth") with
        | true, v ->
            let prot = idp.CreateProtector("auth-cookie")
            let v = prot.Unprotect(v)
            match v.Split ':' with
            | [|"usr"; v|] -> v
            | _ -> "invalid user"
        | _ -> "not logged in"
        next()
        })) |> ignore
```

I keep trying to figure out how to do my own middleware in F#. I finally found this in the journal.

https://cloudandmobiledev.wordpress.com/2018/09/14/how-to-add-middleware-to-asp-net-core-2-1-with-f/

Here is also an way.

But I'll just plug it into the Giraffe pipeline at the start.

4:40pm.

///

In the last video, we gave the advice that you should just forget onsite and use 3rd party, and in this video you can see why.
ASP.NET Core Identity would be perfectly fine on its own, but it turns into a mess when you try to combine different authentication schemes.
Maybe it keeps salaried webdevs busy, but as far as we can see, there are only disadvantages to having multiple ways of authenticating an user. You are just doing extra work for no discernible benefit.
We ran into a lot of extra work
We are done with part 4, but it cannot be said that the onsite login is complete. The biggest problem is that it is missing the email verification part. If you tried using this as it, a malicious user could use it to create accounts for emails he doesn't own.

///

No, let me back this up. I am not done yet. I am going to do email logins.

5:35pm. I showed how to do an email login.

5:45pm. https://youtu.be/1Ifvyud7I70
Authentication & Authorization With ASP.NET Core And Giraffe. ASP.NET Core Identity. (Pt. 4)

Here it is.

5:55pm. Posted on Twitter as well.

6pm. Damn, dealing with this was super hard.

I am drained mentally. No way can I do anything more for the day.

I'll leave SPAs for tomorrow.

Let me close here."