"9:05am. Let me chill for a while and then I will start. I need to take a bath today.
As for that error from yesterday, I am going to ditch the React specific library in favor of the regular one.

10:35am. Let me start for the day. Done with the bath.

///

Day two.

We had time to mull over it during the night. We could go even deeper and try to fix yet another error, but instead let us ditch the Msal React library.

Since we will be using Elmish, we do not need those hard-to-understand hooks anyway, and dealing with the regular library will be a lot easier as it just has 3 methods we need to care about: login, acquire token silent and acquire token interactive.

///

///

Going down the wrong path happens often in programming. Our original plan was to use the Msal library, and then we got distracted by the React version of it.

We don't have an obligation to try and fix every weird error we come across.

///

11:10am.

///

This is really nice. We have a working example that we can use as a reference here.

Let us pause just for a bit so we can study it ourselves.

///

11:20am. Let me resume.

> Hmm, where is the Fable Promise? Maybe it got moved since Fable got more Python and Rust backends?

11:35am.

///

It doesn't seem we need to `open Fable.Promise` explicitly. It started working as soon as we added the package.

Great!

///

12:05pm. https://mangadex.org/title/db60fb9f-60db-4cdb-80ba-f0cc114b28ea/dead-or-deadline
> A man named Chen Ji was obsessed with a manhua called "Assassin A" when he was a child, causing him wanting to become a manhua artist. After years of hard work, he finally succeeded in becoming an assassin.

Lol.

Let me have breakfast.

1:10pm. Let me resume. Done with breakfast and chores.

> Also, we think that you are first supposed to try and acquire the token, and only redirect the user if that fails. We'll have to investigate all of that.

1:25pm. > Before we go any deeper into it, let's try running the thing normally.

1:40pm.

> Ah, so based on the description this isn't supposed to do a redirect, but handle its aftermath. Then it is no wonder it is not redirecting.

3:35pm.

///

Watch the URL bar as we refresh in the background.

We'll slow down a bit here to 1x. As you can see, if we refresh the page, it is not going to and fro the Microsoft site. The silent login works!

///

3:40pm. I need to shut down. There is thunder outside.

4:15pm. It is light. Let me resume.

> Let's test this in incognito mode just to see what will happen. We'll comment out the `InteractionRequiredAuthError` case first.

> Ah, since it doesn't have anything in session storage, it does a redirect. Let us try logging out of our account then.

///

The case is annoyingly hard to try out for some reason. Let's just comment it in, and call it a job well done. We think that the case would trigger once the token expires, but who is going to wait for that?

What comes next?

///

4:50pm. Let me just put some markers.

5:15pm.

> We were focusing on passing the access token too much so we neglected to think where it should be passed. Instead of into the model, it would be better if we passed it into the Fable.Remoting API interface.

> Perfect! Now we can move to authenticating the token on the Server side. Let's try running it first though.

> Of course it doesn't work since the server is forbidding all our requests because we have the wrong middleware in it. To fix that we'll plug in the JWT auth middleware.

5:30pm. > No, it doesn't work. Let's open up the server in the debugger and see what we are getting.

> The Authorization header is what is carrying the access token. But we think it might need to be prefixed with "Bearer ". We can add that to the string, but we'd rather see if we can turn that off in the middleware.

6:15pm. Damn it, let me close here. I am feeling strained. I've got the frontend parts done perfectly, but what I need to figure out is how to deal with the Web APIs.

https://github.com/JanBizub/Fable.Msal/blob/main/src/demo/Server/appsettings.json

https://github.com/JanBizub/Fable.Msal/blob/main/src/demo/Client/src/HttpRequests.fs

He is passing in the accessToken with Bearer.

7pm. Let me close here. It is not as simple as I thought it would be.

While it is true that anyone can validate a JWT token, anyone can also create a JWT token. How would the backend know the token is from Azure AD rather than somebody making up a key?

I'll have to do some study after all.

I'll have to erase the last 2.5 mins of the video. So far I am 27.5m in. I'll finish this tomorrow."