"11:20am. I've gotten into it. I actually got all these packages installed. Tomorrow I'll post it on Github and try importing it.
5/5/2023

8:35am. I am up. Let me chill.

9:15am. As soon as the bathroom clears I'll take a bath and then resume from where I left off yesterday.

10:30am. Done with the bath. Let me start.

https://youtu.be/tlI_07ZZImg
Amazing Card Flip Transition / Generator / Davinci Resolve Tutorial / Fusion

The lack of card transitions has really been irking me. Let me just watch this.

https://youtu.be/tlI_07ZZImg?t=149

This is so annoying. I do not want to program DV every time I do this. Is there a way to get it out of the box? Or maybe I should just do this and find a way to save the transitions.

https://youtu.be/tlI_07ZZImg?t=414

Let me stop this here. I'll come back to it later.

///

And it didn't work. We said we'd scavenge, but...
Let's pause here. We'll try cloning the project again and seeing if we can get it to build in isolation.
Either way we'll have to study it.
I guess this is one of the pains of programming in a relatively unpopular language.
Its ecosystem is not as robust as it would be for bigger languages.
All that we do is for the love of functional programming, so here we'll pause and consider our options more carefully yet again.
We spent a lot of time thinking what we'd do with the library, but not so much what we'd do if it turned out to be out of date.

///

Let me cut this out.

11:15am. Ah, maybe it is trying to get Fable.SignalR.Saturn from the Nuget package instead of my own.

11:20am. https://github.com/fsprojects/Paket/issues/

Check out issue 3064.

> I reviewed the documentation again, GitHub dependencies. At the time I somehow thought that paket allowed the import of github hosted libraries similar to how it treated nuget. Upon review, that is not the case. The documentation in this area could use some work.

11:35am. Holy shit, what are all these errors?

Ah, it must have been caused due to reloading errors.

12:55pm. No, I need to pause and study the library properly without necessarily being tied by the presentation work.

Oh, I see how we could get the state!

Using services.

Still, I do not want to add middleware for every model type. I wonder if it would possible to avoid doing that.

Something like a scoped service for every request, but it is generic object?

https://learn.microsoft.com/en-us/aspnet/signalr/overview/guide-to-the-api/hubs-api-guide-server#hub-object-lifetime

///

Hub object lifetime
You don't instantiate the Hub class or call its methods from your own code on the server; all that is done for you by the SignalR Hubs pipeline. SignalR creates a new instance of your Hub class each time it needs to handle a Hub operation such as when a client connects, disconnects, or makes a method call to the server.

Because instances of the Hub class are transient, you can't use them to maintain state from one method call to the next. Each time the server receives a method call from a client, a new instance of your Hub class processes the message. To maintain state through multiple connections and method calls, use some other method such as a database, or a static variable on the Hub class, or a different class that does not derive from Hub. If you persist data in memory, using a method such as a static variable on the Hub class, the data will be lost when the app domain recycles.

If you want to send messages to clients from your own code that runs outside the Hub class, you can't do it by instantiating a Hub class instance, but you can do it by getting a reference to the SignalR context object for your Hub class. For more information, see How to call client methods and manage groups from outside the Hub class later in this topic.

///

Forget this. Let me chill.

2:40pm. Let me resume. I finally installed the library, and what I have to do is figure out how to deal with state. One method that sticks out is to instantiate a mailbox processor for every request. We'll see...

But I do want to get some measure of control over this.

And that library has little documentation.

Maybe I'd have been better off with raw SignalR + my own serialization.

```fs
hubContext.Context.Items
```

> Gets a key/value collection that can be used to share data within the scope of this connection.

Oh wow. The context also has what I need for aborting the connection. It also has a cancellation token that I can use to detect when the connection is over.

Wonderful.

3pm. https://learn.microsoft.com/en-us/aspnet/webhooks/

Oh webhooks are a thing.

https://learn.microsoft.com/en-us/aspnet/signalr/videos/

Oh there are some vids in the docs.

...A single one.

Ok, focus me.

4:10pm.

```
Type constraint mismatch. The type 'SignalR.State.Send<'f,'g>' is not compatible with type 'SignalR.State.Settings<'a,'b,'c,'d,'e>'
```

Damn it. A really weird error is happening.

```
            configure_signalr {
                endpoint Shared.Constants.endpoint
                send (failwith "")
                invoke (failwith "")
            }
```

Damn it, I need to configure both send and invoke.

4:20pm. Let me watch a few tutorials.

https://github.com/raw-coding-youtube/aspnetcore-signalr/blob/main/3.FeatureOverview/FeatureOverview/Startup.cs

Damn it, I really need this guy's tutorials to remind me of how SignalR should be used.

https://youtu.be/sdsCzyTtp6U?list=PLThyvG1mlMzltDxuQj0uQw1TDu1gJUNeG&t=124

Wait, the hub can see the clients. What about invokes?

https://youtu.be/q5ZHAUUAlQE?list=PLOeFnOV9YBa7nzzuXnThdfsyY06AuCP5V
ASP.NET Core SignalR - Feature Overview

Let me watch it.

https://youtu.be/q5ZHAUUAlQE?list=PLOeFnOV9YBa7nzzuXnThdfsyY06AuCP5V&t=117

On the client side it is using sends.

https://github.com/raw-coding-youtube/aspnetcore-signalr/search?q=invoke

Ah, he does use invoke in the third video.

I am just really confused as to why I am getting passed that untyped FableHub in invoke.

How does it even make a distinction between invoke and send on the server side?

https://youtu.be/q5ZHAUUAlQE?list=PLOeFnOV9YBa7nzzuXnThdfsyY06AuCP5V&t=277

Hmmm, it does seem like it has automatic serialization.

https://youtu.be/q5ZHAUUAlQE?list=PLOeFnOV9YBa7nzzuXnThdfsyY06AuCP5V&t=1099

The context is just the same as the one being passed into the SignalR library.

4:50pm. https://youtu.be/q5ZHAUUAlQE?list=PLOeFnOV9YBa7nzzuXnThdfsyY06AuCP5V&t=1241

Enough, I am thinking about it way too much.

///

Now admittedly, maybe if we tried doing this in vanilla SignalR, it would throw a surprise exception in our face, but...
Hmmm...
Actually, that might very well be what would happen.
You know what, we'll actually try it and report back with the results, stay tuned.

///

Let's give it a try.

https://github.com/raw-coding-youtube/aspnetcore-signalr

I'll clone this guy's repo.

6:35pm. Had to take a break. I really am getting bogged down in this project.

Let me clone that thing so I can try out the invokes.

6:45pm. It doesn't throw an exception, it fucking works.

```cs
        // -------------------------------
        // on connected/disconnected hooks
        // -------------------------------
        public override Task OnConnectedAsync()
        {
            return base.OnConnectedAsync();
        }

        public override Task OnDisconnectedAsync(Exception exception)
        {
            return base.OnDisconnectedAsync(exception);
        }
```

We're really missing these kind of hooks as well.

7:15pm. I'll close it here, I had enough for one day.

I am going to ditch this library in favor of my own stuff.

7:25pm. I did well enough for the day. The recording is now 45m long."